Refactor test-folder
