+++ examples/playground/actix.rs
@@ -123,7 +123,7 @@ fn parse_cmdline<'a>() -> clap::ArgMatches<'a> {
                 .help(
                     &[
                         "Logs `info` in addition to `warn` and `error`.",
-                        "The env-variable `RUST_LOG` has precedence."
+                        "The env-variable `RUST_LOG` has precedence.",
                     ]
                     .join("\n"),
                 ),
