+++ tests/parsing/pbf/mod.rs
@@ -1,0 +1,1 @@
+mod isle_of_man;
